# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Go Loose API",
    "version": "0.1.0",
    "description": "Multi-tenant API-key management and runtime authorization."
  },
  "servers": [{"url": "/"}],
  "components": {
    "securitySchemes": {
      "session": {"type": "apiKey", "in": "cookie", "name": "go_loose_session"},
      "apiKey": {"type": "apiKey", "in": "header", "name": "X-API-Key"},
      "clientBasic": {"type": "http", "scheme": "basic"},
      "userBearer": {"type": "http", "scheme": "bearer"}
    },
...
  },
  "paths": {
...
    "/connect/authorize": {
      "get": {
        "tags": [
          "User Action"
        ],
        "summary": "Start hosted client application login",
        "parameters": [
          {"in": "query", "name": "response_type", "required": true, "schema": {"type": "string", "const": "code"}},
          {"in": "query", "name": "client_id", "required": true, "schema": {"type": "string"}},
          {"in": "query", "name": "redirect_uri", "required": true, "schema": {"type": "string", "format": "uri"}},
          {"in": "query", "name": "state", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {"302": {"description": "Redirect to login or exact client callback"}}
      }
    },
...
  }
}